+++ doc/urls.py
@@ -10,16 +10,15 @@ from . import views
 mo_url_patterns = [
     url(r'^list/', views.DocMOListView.as_view(), name='list', kwargs={'path2': 'List'}),
 
-    url(r'^deploy/', include([
-        url(r'^exec-cmd/$', views.DocDeployExecCmdView.as_view(), name='exec-cmd'),
-        url(r'^minion-list/$', views.DocDeployMinionListView.as_view(), name='minion-list'),
-        url(r'^ssh/', views.DocSSHView.as_view(), name='ssh'),
-        url(r'^sls/', views.DocSLSView.as_view(), name='sls'),
-    ], namespace='deploy'), kwargs={'path2': 'Deploy'}),
+    url(r'^cm/', include([
+        url(r'^exec-cmd/$', views.DocCMExecCmdView.as_view(), name='exec-cmd'),
+        url(r'^minion-list/$', views.DocCMMinionListView.as_view(), name='minion-list'),
+        url(r'^ssh/', views.DocCMSSHView.as_view(), name='ssh'),
+        url(r'^sls/', views.DocCMSLSView.as_view(), name='sls'),
+        url(r'^file-upload/$', views.DocCMFileUploadView.as_view(), name='file-upload'),
+    ], namespace='cm'), kwargs={'path2': 'Cluster Management'}),
 
-    url(r'^file/', include([
-        url(r'^upload/$', views.DocSLSView.as_view(), name='upload'),
-    ], namespace='file'), kwargs={'path2': 'File'}),
+    url(r'^terminal/', views.DocTerminalView.as_view(), name='terminal', kwargs={'path2': 'Terminal'}),
 ]
 
 urlpatterns = [