+++ static/js/mo/salt-minion-list.js
@@ -27,22 +27,32 @@ function refreshMinionsHealth(ele, url) {
 }
 
 // ajax: accept minion key
-function acceptMinion() {
-    var url = $(this).parent('td').siblings('input')[0].value
+function acceptMinion(ele) {
+    var url = $(ele).parent('td').siblings('input')[0].value
     $.ajax({
         url: url,
         beforeSend: function (xhr) {
-            $(this).button('loading');
+            $(ele).button('loading');
         },
         method: 'PUT'
     }).done(function (data, status, xhr) {
         if (xhr.status == 200) {
             alert('Accept Success.')
-            $(this).button('reset')
-            window.location.reload()
         }
     }).fail(function (err) {
-        alert('Oops, something wrong, please contact your Administrator.')
-        $(this).button('reset')
+        if (err.status == 400) {
+            alert('Oops, params error.')
+        } else if (err.status == 500){
+            alert('Oops, something wrong, please contact your Administrator.')
+        }
+    }).always(function () {
+        $(ele).button('reset')
+        window.location.reload()
     })
+}
+
+// delete minion
+function delMinion(ele, url, hostname) {
+    let modalBody = `<p>Hostname: ${hostname}</p>`
+    del(ele, url, modalBody)
 }