+++ terminal/config.py
@@ -1,0 +1,5 @@
+from tornado.options import define
+
+
+def init_config():
+    define('port', default=9527, type=int, help='server listening port')