+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## version 1.4 release (2018.02.08)
+
+1. Fix bugs.
+2. Complete *Terminal* module.
+3. Add *Terminal* doc.
+4. Update Arch.
+
+## version 1.3.4 (2018.02.08)
+
+1. Fix bugs.
+2. `terminal:list` rc.
+
+## version 1.3.3 (2018.02.07)
+
+1. Fix bugs.
+2. Update all models' id from auto_increment to uuid.
+3. Do `terminal:list`.
+
+## version 1.3.2 (2018.02.07)
+
+1. Fix bugs.
+2. Go through the webssh data.
+
+## version 1.3.1 (2018.02.01)
+
+1. Fix bugs.
+2. Change *Deploy* app to *CM* app.
+
 ## version 1.3 release (2018.01.31)
 
 1. Optimize *File Upload*.